backend/requirements.txt
fastapi
uvicorn==0.24.0
python-multipart
pillow
boto3==1.34.0
python-dotenv==1.0.0